separate downloader into files to ease restructure
