+++ pkg/pillar/cmd/downloader/dirs.go
@@ -1,0 +1,57 @@
+package downloader
+
+import (
+	"os"
+
+	"github.com/lf-edge/eve/pkg/pillar/types"
+	log "github.com/sirupsen/logrus"
+)
+
+func initializeDirs() {
+
+	// Remove any files which didn't make it to the verifier.
+	// XXX space calculation doesn't take into account files in verifier
+	// XXX get space report from verifier??
+	clearInProgressDownloadDirs(downloaderObjTypes)
+
+	// create the object download directories
+	createDownloadDirs(downloaderObjTypes)
+}
+
+// Create the object download directories we own
+func createDownloadDirs(objTypes []string) {
+
+	workingDirTypes := []string{"pending"}
+
+	// now create the download dirs
+	for _, objType := range objTypes {
+		for _, dirType := range workingDirTypes {
+			dirName := types.DownloadDirname + "/" + objType + "/" + dirType
+			if _, err := os.Stat(dirName); err != nil {
+				log.Debugf("Create %s\n", dirName)
+				if err := os.MkdirAll(dirName, 0700); err != nil {
+					log.Fatal(err)
+				}
+			}
+		}
+	}
+}
+
+// clear in-progress object download directories
+func clearInProgressDownloadDirs(objTypes []string) {
+
+	inProgressDirTypes := []string{"pending"}
+
+	// now create the download dirs
+	for _, objType := range objTypes {
+		for _, dirType := range inProgressDirTypes {
+			dirName := types.DownloadDirname + "/" + objType +
+				"/" + dirType
+			if _, err := os.Stat(dirName); err == nil {
+				if err := os.RemoveAll(dirName); err != nil {
+					log.Fatal(err)
+				}
+			}
+		}
+	}
+}
